fix: allow floatable passive numbers

+++ Common/Data/Models/PassiveData.cs
@@ -9,7 +9,7 @@ public class PassiveData()
 	public PassivePosition Position { get; set; }
 	public int MaxLevel { get; set; }
 	public List<PassiveConnection> Connections { get; set; }
-	public int Value { get; set; }
+	public float Value { get; set; }
 	public bool IsHidden { get; set; }
 	public bool IsChoiceNode { get; set; }
 	public int RequiredAllocatedEdges { get; set; } = 1;
